+++ src/shared_allocator.cpp
@@ -1,0 +1,6 @@
+// Chemfiles, a modern library for chemistry file reading and writing
+// Copyright (C) Guillaume Fraux and contributors -- BSD license
+#include "chemfiles/shared_allocator.hpp"
+
+// define the global allocator instance
+shared_allocator shared_allocator::instance_;