+++ tests/capi/doc/chfl_topology/from_frame.c
@@ -7,7 +7,7 @@
 int main() {
     // [example]
     CHFL_FRAME* frame = chfl_frame();
-    CHFL_TOPOLOGY* topology = chfl_topology_from_frame(frame);
+    const CHFL_TOPOLOGY* topology = chfl_topology_from_frame(frame);
 
     if (topology == NULL) {
         /* handle error */